Rename Common.ExplodedGraph to Common.BaseExplodedGraph.
